feat: consolidate thor utils (#5661)
* wip: wip

* feat: more consolidation

* feat: more consolidation

* fix: derp error copy

* feat: getInboundAddressDataForChain

* fix: circular deps

* feat: make it better

* fix: merge

* fix: lint

* fix: tests

---------
