+++ src/components/MultiHopTrade/utils.test.ts
@@ -1,13 +1,11 @@
 import { btcAssetId } from '@shapeshiftoss/caip'
 import { Err, Ok } from '@sniptt/monads'
 import { isTradingActive } from 'components/MultiHopTrade/utils'
-import { getInboundAddressDataForChain } from 'lib/swapper/swappers/ThorchainSwapper/utils/getInboundAddressDataForChain'
 import { SwapperName } from 'lib/swapper/types'
+import { getInboundAddressDataForChain } from 'lib/utils/thorchain/getInboundAddressDataForChain'
 
-jest.mock('lib/swapper/swappers/ThorchainSwapper/utils/getInboundAddressDataForChain', () => ({
-  ...jest.requireActual(
-    'lib/swapper/swappers/ThorchainSwapper/utils/getInboundAddressDataForChain',
-  ),
+jest.mock('lib/utils/thorchain/getInboundAddressDataForChain.ts', () => ({
+  ...jest.requireActual('lib/utils/thorchain/getInboundAddressDataForChain.ts'),
   getInboundAddressDataForChain: jest.fn(),
 }))
