+++ src/components/MultiHopTrade/utils.ts
@@ -3,10 +3,10 @@ import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import axios from 'axios'
 import { getConfig } from 'config'
-import { getInboundAddressDataForChain } from 'lib/swapper/swappers/ThorchainSwapper/utils/getInboundAddressDataForChain'
 import { isRune } from 'lib/swapper/swappers/ThorchainSwapper/utils/isRune/isRune'
 import type { SwapErrorRight } from 'lib/swapper/types'
 import { SwapperName } from 'lib/swapper/types'
+import { getInboundAddressDataForChain } from 'lib/utils/thorchain/getInboundAddressDataForChain'
 
 export const isTradingActive = async (
   assetId: AssetId | undefined,
