+++ src/lib/swapper/swappers/ThorchainSwapper/endpoints.ts
@@ -20,6 +20,7 @@ import type {
   TradeQuote,
   UtxoFeeData,
 } from 'lib/swapper/types'
+import { getInboundAddressDataForChain } from 'lib/utils/thorchain/getInboundAddressDataForChain'
 import { assertGetUtxoChainAdapter } from 'lib/utils/utxo'
 import type { AssetsById } from 'state/slices/assetsSlice/assetsSlice'
 
@@ -29,7 +30,6 @@ import type { ThorEvmTradeQuote } from './getThorTradeQuote/getTradeQuote'
 import { getThorTradeQuote } from './getThorTradeQuote/getTradeQuote'
 import { getTradeTxs } from './getTradeTxs/getTradeTxs'
 import { THORCHAIN_AFFILIATE_FEE_BPS } from './utils/constants'
-import { getInboundAddressDataForChain } from './utils/getInboundAddressDataForChain'
 
 const deductOutboundRuneFee = (fee: string): string => {
   // 0.02 RUNE is automatically charged on outbound transactions
