+++ src/lib/swapper/swappers/ThorchainSwapper/evm/utils/getThorTxData.ts
@@ -2,8 +2,8 @@ import { fromAssetId } from '@shapeshiftoss/caip'
 import { getConfig } from 'config'
 import type { Asset } from 'lib/asset-service'
 import { depositWithExpiry } from 'lib/swapper/swappers/ThorchainSwapper/evm/routerCalldata'
-import { getInboundAddressDataForChain } from 'lib/swapper/swappers/ThorchainSwapper/utils/getInboundAddressDataForChain'
 import { isNativeEvmAsset } from 'lib/swapper/swappers/utils/helpers/helpers'
+import { getInboundAddressDataForChain } from 'lib/utils/thorchain/getInboundAddressDataForChain'
 
 type GetEvmThorTxInfoArgs = {
   sellAsset: Asset
