+++ src/lib/utils/thorchain/constants.ts
@@ -1,0 +1,3 @@
+export const THOR_PRECISION = 8
+export const BASE_BPS_POINTS = '10000'
+export const THORCHAIN_AFFILIATE_NAME = 'ss'
