+++ src/lib/utils/thorchain/index.ts
@@ -1,33 +1,47 @@
 import type { AccountId } from '@shapeshiftoss/caip'
+import { type AssetId, bchChainId, fromAccountId, fromAssetId } from '@shapeshiftoss/caip'
+import type { UtxoBaseAdapter, UtxoChainId } from '@shapeshiftoss/chain-adapters'
+import type { HDWallet } from '@shapeshiftoss/hdwallet-core'
 import { TxStatus } from '@shapeshiftoss/unchained-client'
 import type { QueryClient } from '@tanstack/react-query'
-import { queryClient } from 'context/QueryClientProvider/queryClient'
+import axios from 'axios'
+import { getConfig } from 'config'
+import memoize from 'lodash/memoize'
+import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import type { Asset } from 'lib/asset-service'
-import { bnOrZero } from 'lib/bignumber/bignumber'
-import { fromBaseUnit, toBaseUnit } from 'lib/math'
+import type { BigNumber, BN } from 'lib/bignumber/bignumber'
+import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { poll } from 'lib/poll/poll'
-import {
-  type EstimatedFeesQueryKey,
-  queryFn as getEstimatedFeesQueryFn,
-} from 'pages/Lending/hooks/useGetEstimatedFeesQuery'
-import type { IsSweepNeededQueryKey } from 'pages/Lending/hooks/useIsSweepNeededQuery'
-import { queryFn as isSweepNeededQueryFn } from 'pages/Lending/hooks/useIsSweepNeededQuery'
-import { selectPortfolioCryptoBalanceBaseUnitByFilter } from 'state/slices/common-selectors'
-import type { ThorchainSaversWithdrawQuoteResponseSuccess } from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/types'
-import {
-  fromThorBaseUnit,
-  getThorchainTransactionStatus,
-} from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils'
-import { isUtxoChainId } from 'state/slices/portfolioSlice/utils'
-import { selectMarketDataById } from 'state/slices/selectors'
-import { store } from 'state/store'
+import type { ThornodeStatusResponse } from 'lib/swapper/swappers/ThorchainSwapper/types'
+import type { getThorchainSaversPosition } from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils'
+import type { AccountMetadata } from 'state/slices/portfolioSlice/portfolioSliceCommon'
+import { isUtxoAccountId, isUtxoChainId } from 'state/slices/portfolioSlice/utils'
 
-import type { GetThorchainSaversDepositQuoteQueryKey } from './hooks/useGetThorchainSaversDepositQuoteQuery'
-import { queryFn as getThorchainSaversDepositQuoteQueryFn } from './hooks/useGetThorchainSaversDepositQuoteQuery'
-import {
-  type GetThorchainSaversWithdrawQuoteQueryKey,
-  queryFn as getThorchainSaversWithdrawQuoteQueryFn,
-} from './hooks/useGetThorchainSaversWithdrawQuoteQuery'
+import { THOR_PRECISION } from './constants'
+import type { getThorchainLendingPosition } from './lending'
+
+const getThorchainTransactionStatus = async (txHash: string, skipOutbound?: boolean) => {
+  const thorTxHash = txHash.replace(/^0x/, '')
+  const { data: thorTxData, status } = await axios.get<ThornodeStatusResponse>(
+    `${getConfig().REACT_APP_THORCHAIN_NODE_URL}/lcd/thorchain/tx/status/${thorTxHash}`,
+    // We don't want to throw on 404s, we're parsing these ourselves
+    { validateStatus: () => true },
+  )
+
+  if ('error' in thorTxData || status === 404) return TxStatus.Unknown
+  // Tx has been observed, but swap/outbound Tx hasn't been completed yet
+  if (
+    // Despite the Tx being observed, things may be slow to be picked on the THOR node side of things i.e for swaps to/from BTC
+    thorTxData.stages.inbound_finalised?.completed === false ||
+    thorTxData.stages.swap_status?.pending === true ||
+    (!skipOutbound && thorTxData.stages.outbound_signed?.completed === false)
+  )
+    return TxStatus.Pending
+  if (thorTxData.stages.swap_status?.pending === false) return TxStatus.Confirmed
+
+  // We shouldn't end up here, but just in case
+  return TxStatus.Unknown
+}
 
 export const waitForThorchainUpdate = ({
   txHash,
@@ -49,208 +63,100 @@ export const waitForThorchainUpdate = ({
     maxAttempts: 20,
   })
 
-// TODO(gomes): this will work for UTXO but is invalid for tokens since they use diff. denoms
-// the current workaround is to not do fee deduction for non-UTXO chains,
-// but for consistency, we should for native EVM assets, and ensure this is a no-op for tokens
-// Note when implementing this, fee checks/deduction will need to either be done for *native* assets only
-// or handle different denoms for tokens/native assets and display insufficientFundsForProtocolFee copy
-const getHasEnoughBalanceForTxPlusFees = ({
-  balanceCryptoBaseUnit,
-  amountCryptoPrecision,
-  txFeeCryptoBaseUnit,
-  precision,
-}: {
-  balanceCryptoBaseUnit: string
-  amountCryptoPrecision: string
-  txFeeCryptoBaseUnit: string
-  precision: number
-}) => {
-  const balanceCryptoBaseUnitBn = bnOrZero(balanceCryptoBaseUnit)
-  if (balanceCryptoBaseUnitBn.isZero()) return false
+export const fromThorBaseUnit = (valueThorBaseUnit: BigNumber.Value | null | undefined): BN =>
+  bnOrZero(valueThorBaseUnit).div(bn(10).pow(THOR_PRECISION)) // to crypto precision from THOR 8 dp base unit
 
-  return bnOrZero(amountCryptoPrecision)
-    .plus(fromBaseUnit(txFeeCryptoBaseUnit, precision))
-    .lte(fromBaseUnit(balanceCryptoBaseUnitBn, precision))
-}
-
-const getHasEnoughBalanceForTxPlusFeesPlusSweep = ({
-  balanceCryptoBaseUnit,
-  amountCryptoPrecision,
-  txFeeCryptoBaseUnit,
-  precision,
-  sweepTxFeeCryptoBaseUnit,
-}: {
-  balanceCryptoBaseUnit: string
-  amountCryptoPrecision: string
-  txFeeCryptoBaseUnit: string
-  precision: number
-  sweepTxFeeCryptoBaseUnit: string
-}) => {
-  const balanceCryptoBaseUnitBn = bnOrZero(balanceCryptoBaseUnit)
-  if (balanceCryptoBaseUnitBn.isZero()) return { hasEnoughBalance: false, missingFunds: null }
-
-  return {
-    hasEnoughBalance: bnOrZero(amountCryptoPrecision)
-      .plus(fromBaseUnit(txFeeCryptoBaseUnit, precision))
-      .plus(fromBaseUnit(sweepTxFeeCryptoBaseUnit, precision))
-      .lte(fromBaseUnit(balanceCryptoBaseUnitBn, precision)),
-    missingFunds: bnOrZero(amountCryptoPrecision)
-      .plus(fromBaseUnit(txFeeCryptoBaseUnit, precision))
-      .plus(fromBaseUnit(sweepTxFeeCryptoBaseUnit, precision))
-      .minus(fromBaseUnit(balanceCryptoBaseUnitBn, precision)),
-  }
-}
-
-export const fetchHasEnoughBalanceForTxPlusFeesPlusSweep = async ({
-  amountCryptoPrecision: _amountCryptoPrecision,
-  accountId,
+export const toThorBaseUnit = ({
+  valueCryptoBaseUnit,
   asset,
-  type,
-  fromAddress,
 }: {
+  valueCryptoBaseUnit: BigNumber.Value | null | undefined
   asset: Asset
-  fromAddress: string | null
-  amountCryptoPrecision: string
-  accountId: AccountId
-  type: 'deposit' | 'withdraw'
-}) => {
-  const isUtxoChain = isUtxoChainId(asset.chainId)
-  const estimateFeesQueryEnabled = Boolean(fromAddress && accountId && isUtxoChain)
+}): BN => {
+  if (!asset?.precision) return bn(0)
 
-  const balanceCryptoBaseUnit = selectPortfolioCryptoBalanceBaseUnitByFilter(store.getState(), {
-    assetId: asset.assetId,
-    accountId,
-  })
-  const assetMarketData = selectMarketDataById(store.getState(), asset.assetId)
-  const quote = await (async () => {
-    switch (type) {
-      case 'withdraw': {
-        const withdrawAmountCryptoBaseUnit = toBaseUnit(_amountCryptoPrecision, asset.precision)
-
-        const thorchainSaversWithdrawQuoteQueryKey: GetThorchainSaversWithdrawQuoteQueryKey = [
-          'thorchainSaversWithdrawQuote',
-          { asset, accountId, amountCryptoBaseUnit: withdrawAmountCryptoBaseUnit },
-        ]
-
-        return queryClient.fetchQuery({
-          queryKey: thorchainSaversWithdrawQuoteQueryKey,
-          queryFn: getThorchainSaversWithdrawQuoteQueryFn,
-          staleTime: 5000,
-        })
-      }
-      case 'deposit': {
-        const amountCryptoBaseUnit = toBaseUnit(_amountCryptoPrecision, asset.precision)
-
-        const thorchainSaversDepositQuoteQueryKey: GetThorchainSaversDepositQuoteQueryKey = [
-          'thorchainSaversDepositQuote',
-          { asset, amountCryptoBaseUnit },
-        ]
-
-        return await queryClient.fetchQuery({
-          queryKey: thorchainSaversDepositQuoteQueryKey,
-          queryFn: getThorchainSaversDepositQuoteQueryFn,
-          staleTime: 5000,
-        })
-      }
-      default:
-        throw new Error('Invalid type')
-    }
-  })()
-  const amountCryptoPrecision =
-    type === 'deposit'
-      ? _amountCryptoPrecision
-      : fromThorBaseUnit(
-          (quote as ThorchainSaversWithdrawQuoteResponseSuccess).dust_amount,
-        ).toFixed()
-  const amountCryptoBaseUnit = toBaseUnit(amountCryptoPrecision, asset.precision)
-  const estimatedFeesQueryArgs = {
-    estimateFeesInput: {
-      cryptoAmount: amountCryptoPrecision,
-      assetId: asset.assetId,
-      to: quote?.inbound_address ?? '',
-      sendMax: false,
-      accountId: accountId ?? '',
-      contractAddress: undefined,
-    },
-    asset,
-    assetMarketData,
-    enabled: estimateFeesQueryEnabled,
-  }
-
-  const estimatedFeesQueryKey: EstimatedFeesQueryKey = ['estimateFees', estimatedFeesQueryArgs]
-
-  const _estimatedFeesData = estimateFeesQueryEnabled
-    ? await queryClient.fetchQuery({
-        queryKey: estimatedFeesQueryKey,
-        queryFn: getEstimatedFeesQueryFn,
-      })
-    : undefined
-
-  const _hasEnoughBalanceForTxPlusFees = getHasEnoughBalanceForTxPlusFees({
-    precision: asset.precision,
-    balanceCryptoBaseUnit,
-    amountCryptoPrecision,
-    txFeeCryptoBaseUnit: _estimatedFeesData?.txFeeCryptoBaseUnit ?? '',
-  })
-
-  const isSweepNeededQueryEnabled = Boolean(
-    bnOrZero(amountCryptoPrecision).gt(0) && _estimatedFeesData && _hasEnoughBalanceForTxPlusFees,
-  )
-
-  const isSweepNeededQueryArgs = {
-    assetId: asset.assetId,
-    address: fromAddress,
-    txFeeCryptoBaseUnit: _estimatedFeesData?.txFeeCryptoBaseUnit ?? '0',
-    amountCryptoBaseUnit,
-  }
-
-  const isSweepNeededQueryKey: IsSweepNeededQueryKey = ['isSweepNeeded', isSweepNeededQueryArgs]
-
-  const _isSweepNeeded = isSweepNeededQueryEnabled
-    ? await queryClient.fetchQuery({
-        queryKey: isSweepNeededQueryKey,
-        queryFn: isSweepNeededQueryFn,
-      })
-    : undefined
-
-  const isEstimateSweepFeesQueryEnabled = Boolean(_isSweepNeeded && accountId && isUtxoChain)
-
-  const estimatedSweepFeesQueryArgs = {
-    asset,
-    assetMarketData,
-    estimateFeesInput: {
-      cryptoAmount: '0',
-      assetId: asset.assetId,
-      to: fromAddress ?? '',
-      sendMax: true,
-      accountId: accountId ?? '',
-      contractAddress: undefined,
-    },
-    enabled: isEstimateSweepFeesQueryEnabled,
-  }
-
-  const estimatedSweepFeesQueryKey: EstimatedFeesQueryKey = [
-    'estimateFees',
-    estimatedSweepFeesQueryArgs,
-  ]
-
-  const _estimatedSweepFeesData = isEstimateSweepFeesQueryEnabled
-    ? await queryClient.fetchQuery({
-        queryKey: estimatedSweepFeesQueryKey,
-        queryFn: getEstimatedFeesQueryFn,
-      })
-    : undefined
-
-  const { hasEnoughBalance, missingFunds } = getHasEnoughBalanceForTxPlusFeesPlusSweep({
-    precision: asset.precision,
-    balanceCryptoBaseUnit,
-    amountCryptoPrecision,
-    txFeeCryptoBaseUnit: _estimatedFeesData?.txFeeCryptoBaseUnit ?? '0',
-    sweepTxFeeCryptoBaseUnit: _estimatedSweepFeesData?.txFeeCryptoBaseUnit ?? '0',
-  })
-
-  const hasEnoughBalanceForTxPlusFeesPlusSweep =
-    bnOrZero(balanceCryptoBaseUnit).gt(0) && hasEnoughBalance
-
-  return { hasEnoughBalance: hasEnoughBalanceForTxPlusFeesPlusSweep, missingFunds }
+  return bnOrZero(valueCryptoBaseUnit)
+    .div(bn(10).pow(asset?.precision)) // to crypto precision from THOR 8 dp base unit
+    .times(bn(10).pow(THOR_PRECISION))
+    .decimalPlaces(0) // THORChain expects ints, not floats
 }
+
+export const getThorchainFromAddress = async ({
+  accountId,
+  assetId,
+  getPosition,
+  accountMetadata,
+  wallet,
+}: {
+  accountId: AccountId
+  assetId: AssetId
+  getPosition: typeof getThorchainLendingPosition | typeof getThorchainSaversPosition
+  accountMetadata: AccountMetadata
+  wallet: HDWallet
+}): Promise<string> => {
+  const { chainId } = fromAssetId(assetId)
+  if (!isUtxoChainId(chainId)) return Promise.resolve(fromAccountId(accountId).account)
+
+  try {
+    const position = await getPosition({
+      accountId,
+      assetId,
+    })
+    if (!position) throw new Error(`No position found for assetId: ${assetId}`)
+    const address: string = (() => {
+      // THORChain lending position
+      if ('owner' in position) return position.owner
+      // THORChain savers position
+      if ('asset_address' in position) return position.asset_address
+      // For type completeness - if we have a response, we *should* either have an `owner` or `asset_address` property
+      return ''
+    })()
+    return chainId === bchChainId ? `bitcoincash:${address}` : address
+  } catch {
+    const accountType = accountMetadata?.accountType
+    const bip44Params = accountMetadata?.bip44Params
+
+    const chainAdapter = getChainAdapterManager().get(chainId)!
+
+    const firstReceiveAddress = await chainAdapter.getAddress({
+      wallet,
+      accountNumber: bip44Params.accountNumber,
+      accountType,
+      index: 0,
+    })
+    return firstReceiveAddress
+  }
+}
+
+const getAccountAddressesWithBalances = async (
+  accountId: AccountId,
+): Promise<{ address: string; balance: string }[]> => {
+  if (isUtxoAccountId(accountId)) {
+    const { chainId, account: pubkey } = fromAccountId(accountId)
+    const chainAdapters = getChainAdapterManager()
+    const adapter = chainAdapters.get(chainId) as unknown as UtxoBaseAdapter<UtxoChainId>
+    if (!adapter) throw new Error(`no adapter for ${chainId} not available`)
+
+    const {
+      chainSpecific: { addresses },
+    } = await adapter.getAccount(pubkey)
+
+    if (!addresses) return []
+
+    return addresses.map(({ pubkey, balance }) => {
+      const address = pubkey.startsWith('bitcoincash') ? pubkey.replace('bitcoincash:', '') : pubkey
+
+      return { address, balance }
+    })
+  }
+
+  // We don't need balances for chain others than UTXOs
+  return [{ address: fromAccountId(accountId).account, balance: '' }]
+}
+
+// Memoized on accountId, see lodash docs:
+// "By default, the first argument provided to the memoized function is used as the map cache key."
+export const getAccountAddresses = memoize(
+  async (accountId: AccountId): Promise<string[]> =>
+    (await getAccountAddressesWithBalances(accountId)).map(({ address }) => address),
+)
