+++ src/state/slices/opportunitiesSlice/resolvers/thorchainsavers/types.ts
@@ -1,5 +1,6 @@
+import type { QuoteFees } from 'lib/utils/thorchain/lending/types'
+
 import type { DefiProvider, DefiType, OpportunityMetadataBase } from '../../types'
-import type { QuoteFees } from '../thorchainLending/types'
 
 export type ThorchainSaversStakingSpecificMetadata = OpportunityMetadataBase & {
   provider: DefiProvider.ThorchainSavers
