+++ src/state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils.ts
@@ -2,38 +2,26 @@ import type { AccountId, AssetId, ChainId } from '@shapeshiftoss/caip'
 import {
   avalancheAssetId,
   bchAssetId,
-  bchChainId,
   binanceAssetId,
   bscAssetId,
   btcAssetId,
   cosmosAssetId,
   dogeAssetId,
   ethAssetId,
-  fromAccountId,
   fromAssetId,
   ltcAssetId,
 } from '@shapeshiftoss/caip'
-import type { UtxoBaseAdapter, UtxoChainId } from '@shapeshiftoss/chain-adapters'
-import type { HDWallet } from '@shapeshiftoss/hdwallet-core'
-import { TxStatus } from '@shapeshiftoss/unchained-client'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import axios from 'axios'
 import { getConfig } from 'config'
-import memoize from 'lodash/memoize'
-import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import type { Asset } from 'lib/asset-service'
-import type { BN } from 'lib/bignumber/bignumber'
-import { BigNumber, bn, bnOrZero } from 'lib/bignumber/bignumber'
-import type {
-  ThornodePoolResponse,
-  ThornodeStatusResponse,
-} from 'lib/swapper/swappers/ThorchainSwapper/types'
+import { BigNumber, bnOrZero } from 'lib/bignumber/bignumber'
+import type { ThornodePoolResponse } from 'lib/swapper/swappers/ThorchainSwapper/types'
 import { assetIdToPoolAssetId } from 'lib/swapper/swappers/ThorchainSwapper/utils/poolAssetHelpers/poolAssetHelpers'
-import type { AccountMetadata } from 'state/slices/portfolioSlice/portfolioSliceCommon'
-import { isUtxoAccountId, isUtxoChainId } from 'state/slices/portfolioSlice/utils'
+import { fromThorBaseUnit, getAccountAddresses, toThorBaseUnit } from 'lib/utils/thorchain'
+import { BASE_BPS_POINTS, THORCHAIN_AFFILIATE_NAME } from 'lib/utils/thorchain/constants'
 
-import type { getThorchainLendingPosition } from '../thorchainLending/utils'
 import type {
   MidgardPoolPeriod,
   MidgardPoolRequest,
@@ -45,10 +33,6 @@ import type {
   ThorchainSaversWithdrawQuoteResponseSuccess,
 } from './types'
 
-const THOR_PRECISION = '8'
-export const BASE_BPS_POINTS = '10000'
-
-export const THORCHAIN_AFFILIATE_NAME = 'ss'
 // BPS are needed as part of the memo, but 0bps won't incur any fees, only used for tracking purposes for now
 const AFFILIATE_BPS = 0
 
@@ -92,39 +76,6 @@ const SUPPORTED_THORCHAIN_SAVERS_CHAIN_IDS = SUPPORTED_THORCHAIN_SAVERS_ASSET_ID
   assetId => fromAssetId(assetId).chainId,
 )
 
-export const getAccountAddressesWithBalances = async (
-  accountId: AccountId,
-): Promise<{ address: string; balance: string }[]> => {
-  if (isUtxoAccountId(accountId)) {
-    const { chainId, account: pubkey } = fromAccountId(accountId)
-    const chainAdapters = getChainAdapterManager()
-    const adapter = chainAdapters.get(chainId) as unknown as UtxoBaseAdapter<UtxoChainId>
-    if (!adapter) throw new Error(`no adapter for ${chainId} not available`)
-
-    const {
-      chainSpecific: { addresses },
-    } = await adapter.getAccount(pubkey)
-
-    if (!addresses) return []
-
-    return addresses.map(({ pubkey, balance }) => {
-      const address = pubkey.startsWith('bitcoincash') ? pubkey.replace('bitcoincash:', '') : pubkey
-
-      return { address, balance }
-    })
-  }
-
-  // We don't need balances for chain others than UTXOs
-  return [{ address: fromAccountId(accountId).account, balance: '' }]
-}
-
-// Memoized on accountId, see lodash docs:
-// "By default, the first argument provided to the memoized function is used as the map cache key."
-export const getAccountAddresses = memoize(
-  async (accountId: AccountId): Promise<string[]> =>
-    (await getAccountAddressesWithBalances(accountId)).map(({ address }) => address),
-)
-
 export const getThorchainPools = async (): Promise<ThornodePoolResponse[]> => {
   const { data: opportunitiesData } = await axios.get<ThornodePoolResponse[]>(
     `${getConfig().REACT_APP_THORCHAIN_NODE_URL}/lcd/thorchain/pools`,
@@ -149,29 +100,6 @@ export const getAllThorchainSaversPositions = async (
   if (!opportunitiesData) return []
 
   return opportunitiesData
-}
-
-export const getThorchainTransactionStatus = async (txHash: string, skipOutbound?: boolean) => {
-  const thorTxHash = txHash.replace(/^0x/, '')
-  const { data: thorTxData, status } = await axios.get<ThornodeStatusResponse>(
-    `${getConfig().REACT_APP_THORCHAIN_NODE_URL}/lcd/thorchain/tx/status/${thorTxHash}`,
-    // We don't want to throw on 404s, we're parsing these ourselves
-    { validateStatus: () => true },
-  )
-
-  if ('error' in thorTxData || status === 404) return TxStatus.Unknown
-  // Tx has been observed, but swap/outbound Tx hasn't been completed yet
-  if (
-    // Despite the Tx being observed, things may be slow to be picked on the THOR node side of things i.e for swaps to/from BTC
-    thorTxData.stages.inbound_finalised?.completed === false ||
-    thorTxData.stages.swap_status?.pending === true ||
-    (!skipOutbound && thorTxData.stages.outbound_signed?.completed === false)
-  )
-    return TxStatus.Pending
-  if (thorTxData.stages.swap_status?.pending === false) return TxStatus.Confirmed
-
-  // We shouldn't end up here, but just in case
-  return TxStatus.Unknown
 }
 
 export const getThorchainSaversPosition = async ({
@@ -289,24 +217,6 @@ export const getMidgardPools = async (
   return poolsData
 }
 
-export const fromThorBaseUnit = (valueThorBaseUnit: BigNumber.Value | null | undefined): BN =>
-  bnOrZero(valueThorBaseUnit).div(bn(10).pow(THOR_PRECISION)) // to crypto precision from THOR 8 dp base unit
-
-export const toThorBaseUnit = ({
-  valueCryptoBaseUnit,
-  asset,
-}: {
-  valueCryptoBaseUnit: BigNumber.Value | null | undefined
-  asset: Asset
-}): BN => {
-  if (!asset?.precision) return bn(0)
-
-  return bnOrZero(valueCryptoBaseUnit)
-    .div(bn(10).pow(asset?.precision)) // to crypto precision from THOR 8 dp base unit
-    .times(bn(10).pow(THOR_PRECISION))
-    .decimalPlaces(0) // THORChain expects ints, not floats
-}
-
 export const isAboveDepositDustThreshold = ({
   valueCryptoBaseUnit,
   assetId,
@@ -376,50 +286,4 @@ export const makeDaysToBreakEven = ({
   // and ~ a day after deposit
   const daysToBreakEven = BigNumber.max(daysToBreakEvenOrZero, 1).toFixed(0)
   return daysToBreakEven
-}
-
-export const getThorchainFromAddress = async ({
-  accountId,
-  assetId,
-  getPosition,
-  accountMetadata,
-  wallet,
-}: {
-  accountId: AccountId
-  assetId: AssetId
-  getPosition: typeof getThorchainLendingPosition | typeof getThorchainSaversPosition
-  accountMetadata: AccountMetadata
-  wallet: HDWallet
-}): Promise<string> => {
-  const { chainId } = fromAssetId(assetId)
-  if (!isUtxoChainId(chainId)) return Promise.resolve(fromAccountId(accountId).account)
-
-  try {
-    const position = await getPosition({
-      accountId,
-      assetId,
-    })
-    if (!position) throw new Error(`No position found for assetId: ${assetId}`)
-    const address: string = (() => {
-      // THORChain lending position
-      if ('owner' in position) return position.owner
-      // THORChain savers position
-      if ('asset_address' in position) return position.asset_address
-      // For type completeness - if we have a response, we *should* either have an `owner` or `asset_address` property
-      return ''
-    })()
-    return chainId === bchChainId ? `bitcoincash:${address}` : address
-  } catch {
-    const accountType = accountMetadata?.accountType
-    const bip44Params = accountMetadata?.bip44Params
-    const chainAdapter = getChainAdapterManager().get(chainId)!
-
-    const firstReceiveAddress = await chainAdapter.getAddress({
-      wallet,
-      accountNumber: bip44Params.accountNumber,
-      accountType,
-      index: 0,
-    })
-    return firstReceiveAddress
-  }
 }
